+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
 - Most magical consumable items will now be parsed as Wondrous item type rather than Trinket (added in 5.2).
+- Debug generator tweaks.
 
 # 6.6.48