+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next Up
 
+# 6.6.49
+
 - Most magical consumable items will now be parsed as Wondrous item type rather than Trinket (added in 5.2).
 - Debug generator tweaks.